# msg.Service: optimize a bit